Exercise 4.21

+++ part4/controllers/blogs.js
@@ -38,8 +38,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token is invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id) {
+        await Blog.findByIdAndRemove(request.params.id)
+
+        return response.status(204).end()
+    }
+
+    const error = new Error('The blogs only can be deleted by it\'s creator')
+    error.name = 'ValidationError'
+    throw (error)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
